revert myft-common template changes
