+++ myft/templates/follow.html
@@ -1,1 +1,32 @@
-{{>n-ui/myft-common/templates/follow}}
+{{#if @root.flags.myFtApiWrite}}
+<form class="n-myft-ui n-myft-ui--follow" method="POST"
+	data-myft-ui="follow"
+	data-concept-id="{{conceptId}}"
+	action="/__myft/api/core/followed/concept/{{conceptId}}?method=put">
+	<input type="hidden" value="{{name}}" name="name">
+	<input type="hidden" value="{{taxonomy}}" name="taxonomy">
+	<button
+		aria-label="Follow {{name}}"
+		aria-pressed="false"
+		class="n-myft-ui__button
+			{{#variant}} n-myft-ui__button--{{this}}{{/variant}}
+			{{#size}} n-myft-ui__button--{{this}}{{/size}}"
+		data-alternate-label="Unfollow {{name}}"
+		{{#if alternateText}}
+			data-alternate-text="{{alternateText}}"
+		{{else}}
+		{{#if buttonText}}
+			data-alternate-text="{{buttonText}}"
+		{{else}}
+			data-alternate-text="Following"
+		{{/if}}
+		{{/if}}
+		data-concept-id="{{conceptId}}" {{! duplicated here for tracking}}
+		data-trackable="follow"
+		title="Follow {{name}}"
+		type="submit"
+	>{{#if buttonText}}{{buttonText}}{{else}}Follow{{/if}}</button>
+</form>
+{{else}}
+<!-- Follow button hidden due to myFtApiWrite being off -->
+{{/if }}
